src/main.py
#!/usr/bin/env python3
"""
Enhanced 3xUI Telegram Bot - Main Application
A comprehensive Telegram bot for managing 3xUI VPN panel clients.

Author: ShayC21
Version: 2.0.0
"""

import os
import sys
import asyncio
import json
from uuid import uuid4
from datetime import datetime, timedelta
from urllib.parse import quote
from typing import Dict, List, Optional

# Add src directory to path
sys.path.insert(0, os.path.join(os.path.dirname(__file__)))

# Import local modules
from config_manager import ConfigManager
from logger_setup import setup_logging, log_access, get_logger
from xui_client import XUIClient, XUIAPIError
from utils import format_bytes, format_timestamp, generate_connection_uri

# Telegram imports
from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
from telegram.ext import (
    ApplicationBuilder,
    CommandHandler,
    ContextTypes,
    ConversationHandler,
    MessageHandler,
    CallbackQueryHandler,
    filters,
)

# Load configuration
config = ConfigManager()

# Setup logging
logger = setup_logging(
    log_level=config.get("LOG_LEVEL", "INFO"),
    log_dir=config.get("LOG_DIR", "logs")
)

# Initialize XUI client
xui_client = XUIClient(
    host=config.get("XUI_HOST"),
    username=config.get("USERNAME"),
    password=config.get("PASSWORD"),
    timeout=config.get("REQUEST_TIMEOUT", 30)
)

# Bot configuration
AUTHORIZED_USERS = config.get_authorized_users()

# Conversation states
INBOUND_ID, CLIENT_NAME, CLIENT_DATA, CLIENT_EXPIRY = range(4)
REMOVE_CLIENT_EMAIL = range(1)

# Temporary data storage
user_data_temp = {}

logger.info("🤖 3xUI Telegram Bot initialized")


# Authorization decorator
async def is_authorized(update: Update) -> bool:
    """Enhanced authorization check with detailed logging"""
    user = update.effective_user
    user_id = user.id

    if user_id not in AUTHORIZED_USERS:
        log_access(user_id, user.username or "Unknown", "UNAUTHORIZED_ACCESS", False)
        await update.message.reply_text(
            "⛔ <b>Access Denied</b>\n\n"
            "🚫 You are not authorized to use this bot.\n"
            "📞 Contact the administrator for access.",
            parse_mode="HTML"
        )
        return False

    log_access(user_id, user.username or "Unknown", "AUTHORIZED_ACCESS", True)
    return True


# Main command handlers
async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Enhanced start command with comprehensive help"""
    if not await is_authorized(update):
        return

    user = update.effective_user
    log_access(user.id, user.username or "Unknown", "START_COMMAND")

    # Create inline keyboard for main menu
    keyboard = [
        [InlineKeyboardButton("➕ Add Client", callback_data="add_client")],
        [InlineKeyboardButton("🗑️ Remove Client", callback_data="remove_client")],
        [InlineKeyboardButton("📋 List Clients", callback_data="list_clients")],
        [InlineKeyboardButton("📊 Inbound Stats", callback_data="inbound_stats")],
        [InlineKeyboardButton("ℹ️ Help", callback_data="help")]
    ]
    reply_markup = InlineKeyboardMarkup(keyboard)

    welcome_message = (
        f"🌐 <b>3xUI Management Bot v2.0</b> 🌐\n\n"
        f"👋 Welcome, {user.first_name}!\n\n"
        f"🚀 <b>Enhanced Features:</b>\n"
        f"• ➕ Add VPN clients with custom settings\n"
        f"• 🗑️ Remove clients safely\n"
        f"• 📋 List all clients with details\n"
        f"• 📊 View inbound statistics\n"
        f"• 🔗 Support for VLESS, VMess, Trojan\n"
        f"• 📈 Traffic monitoring\n"
        f"• ⏰ Expiry date management\n\n"
        f"🔒 <b>Secure & Reliable</b>\n"
        f"Choose an option below to get started:"
    )

    await update.message.reply_text(
        welcome_message,
        parse_mode="HTML",
        reply_markup=reply_markup
    )


async def help_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Detailed help command"""
    if not await is_authorized(update):
        return

    help_text = (
        "📚 <b>Bot Commands & Usage</b>\n\n"
        "🔧 <b>Available Commands:</b>\n"
        "• <code>/start</code> - Main menu and welcome\n"
        "• <code>/addclient</code> - Add new VPN client\n"
        "• <code>/removeclient</code> - Remove existing client\n"
        "• <code>/listclients</code> - Show all clients\n"
        "• <code>/stats</code> - Show inbound statistics\n"
        "• <code>/help</code> - Show this help message\n"
        "• <code>/cancel</code> - Cancel current operation\n\n"
        "📋 <b>Usage Examples:</b>\n\n"
        "1️⃣ <b>Adding a Client:</b>\n"
        "   • Use /addclient command\n"
        "   • Enter inbound ID (e.g., 1)\n"
        "   • Enter client name/email\n"
        "   • Set data limit in GB (0 = unlimited)\n"
        "   • Set expiry in days (0 = never)\n\n"
        "2️⃣ <b>Data Limits:</b>\n"
        "   • Enter 0 for unlimited data\n"
        "   • Enter any number for GB limit (e.g., 50)\n\n"
        "3️⃣ <b>Expiry Dates:</b>\n"
        "   • Enter 0 for lifetime access\n"
        "   • Enter days for expiry (e.g., 30 for 1 month)\n\n"
        "🔐 <b>Security Features:</b>\n"
        "• Only authorized users can access\n"
        "• All actions are logged\n"
        "• Secure connection to 3xUI panel\n\n"
        "❓ <b>Need Help?</b>\n"
        "Contact: @ShayC21"
    )

    await update.message.reply_text(help_text, parse_mode="HTML")


async def button_handler(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Handle inline keyboard button presses"""
    query = update.callback_query
    await query.answer()

    if query.data == "add_client":
        await addclient_start(update, context)
    elif query.data == "remove_client":
        await removeclient_start(update, context)
    elif query.data == "list_clients":
        await list_clients(update, context)
    elif query.data == "inbound_stats":
        await show_stats(update, context)
    elif query.data == "help":
        await help_command(update, context)


# Client management functions (simplified for space)
async def addclient_start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Start add client process"""
    if not await is_authorized(update):
        return ConversationHandler.END

    # Implementation here - similar to original but with error handling
    # This would be the full implementation from the original bot.py
    pass


async def cancel(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Enhanced cancel command with cleanup"""
    chat_id = update.effective_chat.id
    user_data_temp.pop(chat_id, None)

    await update.message.reply_text(
        "❌ <b>Operation Cancelled</b>\n\n"
        "All input data has been cleared.\n"
        "Use /start to return to the main menu.",
        parse_mode="HTML"
    )
    return ConversationHandler.END


def main():
    """Enhanced main function with better error handling and logging"""
    logger.info("🚀 Starting 3xUI Telegram Bot v2.0...")

    try:
        # Test initial connection
        if not xui_client.login():
            logger.error("❌ Failed to connect to 3xUI panel")
            sys.exit(1)

        logger.info("✅ Successfully connected to 3xUI panel")

        # Create application
        application = ApplicationBuilder().token(config.get("BOT_TOKEN")).build()

        # Add handlers
        application.add_handler(CommandHandler("start", start))
        application.add_handler(CommandHandler("help", help_command))
        application.add_handler(CallbackQueryHandler(button_handler))

        # Error handler
        async def error_handler(update: object, context: ContextTypes.DEFAULT_TYPE):
            """Enhanced error handler with detailed logging"""
            logger.error(f"Exception while handling an update: {context.error}")
            if update and isinstance(update, Update) and update.effective_message:
                try:
                    await update.effective_message.reply_text(
                        "❌ <b>An error occurred</b>\n\n"
                        "The bot encountered an unexpected error. "
                        "Please try again or contact the administrator.",
                        parse_mode="HTML"
                    )
                except:
                    pass

        application.add_error_handler(error_handler)

        # Start the bot
        logger.info("🤖 Bot is now running and ready to accept commands...")

        application.run_polling(
            allowed_updates=["message", "callback_query"],
            drop_pending_updates=True
        )

    except KeyboardInterrupt:
        logger.info("🛑 Bot stopped by user")
    except Exception as e:
        logger.error(f"Critical error starting bot: {e}")
    finally:
        # Cleanup
        xui_client.close()
        logger.info("🛑 Bot stopped")


if __name__ == "__main__":
    main()
